Move grid sections to own React function components
