+++ src/App.tsx
@@ -1,7 +1,7 @@
 import "bootstrap/dist/css/bootstrap.min.css";
 import "holderjs";
 import "./App.css";
-import MatrixBackground from "./components/MatrixBackground";
+import MatrixBackground from "./components/MatrixBackground/MatrixBackground";
 import NewNote from "./components/NewNote/NewNote";
 import Title from "./components/Title";
