+++ src/components/MatrixBackground/Delegate.tsx
@@ -1,0 +1,5 @@
+export default function Delegate():JSX.Element{
+    return (
+        <div className="grid">Delegate</div>
+    )
+}
