+++ src/components/MatrixBackground/DoFirst.tsx
@@ -1,0 +1,5 @@
+export default function DoFirst():JSX.Element{
+    return (
+        <div className="grid">Do First</div>
+    )
+}
