+++ src/components/MatrixBackground/ForgetAboutIt.tsx
@@ -1,0 +1,10 @@
+export default function ForgetAboutIt():JSX.Element{
+    return (
+        <div className="grid">
+            Forget about it
+            <ul id='forget_about_it'>
+            <li></li>
+            </ul>
+        </div>
+    )
+}
