+++ src/components/MatrixBackground/Schedule.tsx
@@ -1,0 +1,5 @@
+export default function Schedule():JSX.Element{
+    return (
+        <div className="grid">Schedule</div>
+    )
+}
